TMS - UserController Point 2 code fix
