+++ TMS.API/Web.config
@@ -21,7 +21,7 @@
       <logger name="*" minlevel="Info" writeTo="logfile" />
     </rules>
   </nlog>
-  <appSettings configSource="AppConfig.config"/>
+  <appSettings configSource="AppConfig.config" />
   <system.web>
     <compilation debug="true" targetFramework="4.6.1" />
     <httpRuntime targetFramework="4.6.1" />
@@ -107,7 +107,7 @@
       </providers>
     </dependencyResolver>
   </Helper.Model>
-  <connectionStrings configSource="ConnString.config"/>
+  <connectionStrings configSource="ConnString.config" />
   <entityFramework>
     <defaultConnectionFactory type="System.Data.Entity.Infrastructure.LocalDbConnectionFactory, EntityFramework">
       <parameters>
